created a seperate component for Camera
